requirements.txt
bluepy
pycryptodomex
paho-mqtt
git+https://github.com/pybluez/pybluez.git#egg=pybluez
